requirements.txt
geojson==2.5.0
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
imgaug==0.3.0
imutils==0.5.3
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.3
multidict==4.7.3
numpy==1.17.0
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
OSMPythonTools==0.2.6
pandas==0.25.3
Pillow==7.0.0
protobuf==3.11.3
scikit-image==0.16.2
scipy==1.4.1
Shapely==1.7.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
